@@ -14,7 +14,7 @@
     "build:pure": "rimraf dist && gradio cc build --no-generate-docs",
     "lint": "npm-run-all --parallel lint:* && npm run format",
     "format": "prettier --cache --write '**/*.{js,jsx,tsx,ts,svelte,less,md,yaml,yml,json}' --no-error-on-unmatched-pattern",
-    "lint:js": "eslint . --fix  --cache --ext .js,.jsx,.ts,.tsx,.svelte",
+    "lint:js": "eslint . --fix  --cache",
     "lint:ts": "tsc",
     "lint:py": "flake8 . && isort . && yapf --in-place --recursive --exclude=node_modules/ .",
     "lint:style": "stylelint --allow-empty-input --cache --fix \"**/*.{less,css}\" ",
@@ -29,26 +29,26 @@
     "@actions/core": "^1.10.1",
     "@actions/exec": "^1.1.1",
     "@actions/github": "^6.0.0",
-    "@changesets/cli": "^2.27.5",
+    "@changesets/cli": "^2.27.6",
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@manypkg/get-packages": "^2.2.1",
     "@modelscope-studio/changelog": "workspace:*",
     "@modelscope-studio/lint-config": "workspace:*",
-    "eslint": "^8.57.0",
+    "eslint": "^9.5.0",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.7",
     "node-fetch": "^3.3.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.3.2",
     "prettier-plugin-packagejson": "^2.5.0",
-    "prettier-plugin-svelte": "^3.2.4",
+    "prettier-plugin-svelte": "^3.2.5",
     "rimraf": "^5.0.7",
     "stylelint": "^16.6.1",
     "svelte": "^4.2.18",
     "tslib": "^2.6.3",
-    "tsx": "^4.15.5",
-    "typescript": "^5.4.5"
+    "tsx": "^4.15.7",
+    "typescript": "^5.5.2"
   },
   "main_changeset": true
 }
